[Test][Problem][Refactor] Refactor problems to run on parallel
