+++ src/__tests__/problems/pegadito/problem.test.ts
@@ -1,0 +1,5 @@
+import {  simpleProblemTest } from '../lib.test'
+
+describe(`Test solution of 3 y 5`, () => {
+    simpleProblemTest(__dirname)
+});
